# Use parallel build for frameworks