--- a/.gitignore
+++ b/.gitignore
@@ -25,3 +25,9 @@ example/bind/android/gradlew*
 example/bind/android/gradle
 example/bind/android/build
 example/bind/android/app/build
+
+vendor
+build/
+.gitignore
+cmd/gomobile/gomobile
+.vscode/launch.json