Currently the build for different targets and CPU architectures is sequential. Even for a small project, it usually takes about two minutes to build for `-target maccatalyst,ios,iossimulator,macos`, which is very different from our perception of build speed of Golang. 
By adopting this parallel build approach, we can save a lot of time.

Simple visualization (`gomobile bind -target maccatalyst,ios,iossimulator,macos ...`):

**Main Branch**

1. Build Mac Catalyst (amd64)
2. Build Mac Catalyst (arm64)
    - Merge into a fat binary
3. Build iOS (arm64)
4. Build iOS Simulator (amd64)
5. Build iOS Simulator (arm64)
    - Merge into a fat binary
6. Build macOS (amd64)
7. Build macOS (arm64)
    - Merge into a fat binary
8. Create xcframework

**This PR**

1. Build different architectures in parallel
2. Merge different architectures for a single target into a fat binary (this is the limitation of xcodebuild. In theory, we don't need to do so. Reference: https://developer.apple.com/forums/thread/666335?answerId=685927022#685927022)
3. Create xcframework

## Quick Benchmark

Local machine comparison with a minimal demo Golang project (on M1 Max):

```shell
# This PR
gomobile-parallel bind -target  -o   30.98s user 17.43s system 183% cpu 26.400 total

# Main branch
gomobile bind -target maccatalyst,ios,iossimulator,macos -o   25.80s user 13.17s system 38% cpu 1:42.43 total
```